Added square values for eval and fixed stupid bug in search

+++ backend/board.c
@@ -49,7 +49,7 @@ char* board_get_fen(Board* board) {
     return board_to_fen(board);
 }
 
-int board_evaluate_current(Board* board) {
+float board_evaluate_current(Board* board) {
     return evaluate_board(board);
 }
